fix: Override of model request concurrency setting
* fix: Override of model request concurrency setting

Motivation

The model-mesh option of limiting per-model request concurrency is not completely configurable currently via the built-in adapters. The DISABLE_MODEL_CONCURRENCY env var toggles this but when enabled it is hardcoded to 1.

Modifications

Rename DISABLE_MODEL_CONCURRENCY env var to LIMIT_PER_MODEL_CONCURRENCY which is less confusing and change its value to be the actual per-model concurrency. Zero means don't enable limited request concurrency mode.

No one is currently using this env var since is not yet used/exposed by the built-in adapter orchestration logic in the controller. An update to the controller will follow.

Result

Per-model-concurrency can be fully configured for the built-in adapters.

* Update where runtimeVersion is overridden based on server metadata call