fix: Include model_type in loadModel modelKey json passed by the puller (#15)
* fix: Include model_type in loadModel modelKey json passed by the puller

Motivation

When using the puller with non-builtin runtimes, the modelType provided in the Predictor is passed within some json in the modelKey field because the internal modelType field is used for model placement constraints (it's historical/unfortunate and we hope to change that soon).

However, after the recent storage-related refactor this separate model_type attribute is no longer being passed through.

Also, the storage_key and storage_params attributes _are_ passed along by the puller but this doesn't really make sense because they're consumed by the puller and intended to be abstracted away from the runtime.

Modifications

- Ensure the model_type json attribute is passed through
- Remove storage_key, storage_params and bucket attributes from the transformed request
- Update unit tests to cover this behaviour

Result

Non-built-in runtimes have direct access to the predictor's modelType and are shielded from storage parameters that they shouldn't have to know about.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>

* Address @tjohnson31415's review comments

Signed-off-by: Nick Hill <nickhill@us.ibm.com>