fix: Propagate context to puller download tasks
Motivation

For some reason, the loadModel context is not passed via the puller's ProcessLoadModelRequest function to pullman's Pull function. This means that storage downloads won't be cancelled when the corresponding request is cancelled or times out.

Modifications

Add a Context argument to the puller's ProcessLoadModelRequest function, use in the call to pull manager's Pull instead of context.TODO().

Result

In-progress model downloads from shared storage downloads will be cancelled when the corresponding request is cancelled or times out.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>