fix: correct mlserver uri conversion regardless of filenames (#43)
#### Motivation

As documented in [this issue](https://github.com/kserve/modelmesh-runtime-adapter/issues/40), there is currently a bug in the conversion of the `uri` provided in the MLServer `model-settings.json` file if the model files are symlinked before the uri conversion occurs. This PR is intended to fix the issue by ensuring that `model-settings.json` will always be processed first.

#### Modifications

The function `adaptModelLayoutForRuntime` in the MLServer adapter `server.go` was modified to ensure that, if present, the `model-settings.json` file will be placed at the head of the list of files to be processed. This makes sure that the config file will always be processed before any symlinks are created for the model files.

A unit test was also added to ensure that the bug was fixed.

#### Result

Incorrect uri conversion bug no longer occurs even when the model files alphabetically precede the filename`model-settings.json`.


Signed-off-by: xinyitan <xin.yi.tan@sap.com>

Co-authored-by: xinyitan <xin.yi.tan@sap.com>